Reorganise tools
