+++ cyber_crew/tools/vulnerability.py
@@ -1,9 +1,27 @@
-"""Tool to search for exploits."""
+"""Command line tools to expose system vulnerabilities."""
 
 from crewai.tools import BaseTool
 from pydantic import BaseModel, Field
 
 from cyber_crew.tools.utils import run_command
+
+
+class FindSuidsInput(BaseModel):
+    """Input schema for FindSuids."""
+
+    pass
+
+
+class FindSuids(BaseTool):
+    """Tool to find SUID files."""
+
+    name: str = "Find SUIDs"
+    description: str = "Tool to find SUID files."
+    args_schema: type[BaseModel] = FindSuidsInput
+
+    def _run(self) -> str:
+        cmd = "find / -perm -4000 -type f 2>/dev/null"
+        return run_command(cmd)
 
 
 class SearchExploitInput(BaseModel):
