--- cyber_crew/tools/vulnerability/find_suids.py
@@ -1,24 +1,0 @@
-"""Tool to search for vulnerable SUIDs."""
-
-from crewai.tools import BaseTool
-from pydantic import BaseModel
-
-from cyber_crew.tools.utils import run_command
-
-
-class FindSuidsInput(BaseModel):
-    """Input schema for FindSuids."""
-
-    pass
-
-
-class FindSuids(BaseTool):
-    """Tool to find SUID files."""
-
-    name: str = "Find SUIDs"
-    description: str = "Tool to find SUID files."
-    args_schema: type[BaseModel] = FindSuidsInput
-
-    def _run(self) -> str:
-        cmd = "find / -perm -4000 -type f 2>/dev/null"
-        return run_command(cmd)
